refactors tests to import fake clib from own file
